+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+= 1.2 // 2021.08-Aug. =
+📦 NEW: Retrieves & saves ecologi user total impact every three hours
+📦 NEW: Retrieves & saves ecologi user total after purchase/renewal
+
 = 1.1 // 2021.07-Jul.31 =
 🐛 FIX: Ensures EDD is not in test mode when processing purchase
 🐛 FIX: Removes 'test' => true from API request